requirements.txt
python-telegram-bot[job-queue]>=21.0
python-dotenv==1.0.0
aiohttp>=3.10.0
requests==2.31.0
selenium==4.15.0
Pillow>=9.0.0